improve search for search engines
some search engines don't respect robots files. Some of those engines are running into loops and cause a lot of load on the server. Because they saw a new timestamp in links on every new load of the search page when looking at the following links:

`Current topics of the last 1 3 7 days`